[19] Pull out parsing into separate module
